# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "reprep API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/notifications": {
      "get": {
        "summary": "Notification Settings",
        "operationId": "notification_settings_api_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lessons": {
                      "type": "boolean"
                    },
                    "assignments": {
                      "type": "boolean"
                    },
                    "bot_started": {
                      "type": "boolean"
                    },
                    "delivery_enabled": {
                      "type": "boolean"
                    },
                    "deliveries": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer"
                      }
                    },
                    "note": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "lessons",
                    "assignments",
                    "bot_started",
                    "delivery_enabled",
                    "deliveries",
                    "note"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "x-roles": [
          "tutor",
          "learner",
          "guardian"
        ],
        "security": [
          {
            "SessionBearer": []
          }
        ]
      },
      "put": {
        "summary": "Save Notifications",
        "operationId": "save_notifications_api_notifications_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Authorization, role or resource visibility error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Version, state or replay conflict when applicable"
          }
        },
        "x-roles": [
          "tutor",
          "learner",
          "guardian"
        ],
        "security": [
          {
            "SessionBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NotificationInput": {
        "properties": {
          "lessons": {
            "type": "boolean",
            "title": "Lessons",
            "default": false
          },
          "assignments": {
            "type": "boolean",
            "title": "Assignments",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "NotificationInput"
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "reference_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "reference_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
    "securitySchemes": {
      "SessionBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "MaxWebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Max-Bot-Api-Secret"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local demo only; configure approved HTTPS for judging"
    }
  ]
}